Refactored several inspections into an annotator (removed tests but kept the test data); added annotating for keyword arguments repetition.
