+++ src/main/resources/SnakemakeBundle.properties
@@ -22,3 +22,9 @@ INSP.NAME.shadow.settings=Shadow must either be 'shallow', 'full', 'minimal', or
 
 # SnakemakeMultipleShadowSettingsInspection
 INSP.NAME.shadow.multiple.settings=Only one argument after shadow section is permitted.
+
+
+### Annotators ###
+ANN.keyword.argument.repeated=Keyword argument repeated.
+ANN.positional.argument.after.keyword.argument=Positional argument after keyword argument.
+ANN.multiple.execution.sections=Multiple run/shell sections.
