--- src/main/resources/inspectionDescriptions/SmkMultipleExecutionSectionsInspection.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-Only one of those sections can be used in a single Snakemake rule: run/shell/script/wrapper/cwl.
-</body>
-</html>
