--- src/main/resources/inspectionDescriptions/SmkPositionalArgsAfterKeywordArgsInspection.html
@@ -1,7 +1,0 @@
-<html>
-<body>
-Positional arguments cannot be preceded by keyword arguments
-(e.g. 'a=foo', a keyword argument, cannot come before 'b', a positional argument,
-but 'b, a=foo' is a valid option).
-</body>
-</html>
